chore: use module on store
