+++ src/store/index.js
@@ -1,0 +1,11 @@
+import Vue from 'vue';
+import Vuex from 'vuex';
+import feed from './modules/feed';
+
+Vue.use(Vuex);
+
+export default new Vuex.Store({
+  modules: {
+    feed,
+  },
+});
